+++ tests/testthat/test-integration.R
@@ -117,7 +117,6 @@ test_that("can convert date time types", {
   df2 <- bq_table_download(tb, quiet = TRUE)
 
   base <- ISOdatetime(2000, 1, 2, 3, 4, 5.67, tz = "UTC")
-  attr(df$datetime, "tzone") <- attr(base, "tzone")
 
   expect_equal(df, df2, tolerance = 0.67)
   expect_equal(df$datetime, base)
@@ -177,7 +176,7 @@ test_that("can convert geography type", {
   tb <- bq_project_query(bigrquery::bq_test_project(), sql, quiet = TRUE)
   df <- bqs_table_download(tb, bigrquery::bq_test_project(), as_tibble = TRUE, quiet = TRUE)
 
-  expect_identical(wk::wkt(df$geography), wk::wkt("POINT(30 10)"))
+  expect_identical(df$geography, wk::wkt("POINT(30 10)"))
 })
 
 test_that("can convert bytes type", {
@@ -188,12 +187,12 @@ test_that("can convert bytes type", {
   df <- bqs_table_download(tb, bigrquery::bq_test_project(), as_tibble = TRUE, quiet = TRUE)
 
   expect_identical(
-    unlist(df$bytes),
-    as.raw(c(
+    df$bytes,
+    blob::as_blob(as.raw(c(
       0x01, 0x01, 0x00, 0x00, 0x00, 0xff, 0xff, 0xff, 0xff,
       0xff, 0xff, 0x3d, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x24,
       0x40
-    ))
+    )))
   )
 })
 
@@ -214,4 +213,36 @@ test_that("nested list type", {
 	expect_equal(df[["aos"]], list(tibble::tibble(a = 1:3, b = c("a", "b", "c"))))
 	expect_equal(df[["soa"]], tibble::tibble(a = list(1:3), b = list(c("a", "b"))))
 
+})
+
+test_that("post process parse works", {
+	auth_fn()
+	sql <- "
+	SELECT
+    '\U0001f603' as unicode,
+    datetime,
+    TRUE as logicaltrue,
+    FALSE as logicalfalse,
+    CAST ('Hi' as BYTES) as bytes,
+    CAST (datetime as DATE) as date,
+    CAST (datetime as TIME) as time,
+    CAST (datetime as TIMESTAMP) as timestamp,
+    ST_GEOGFROMTEXT('POINT (30 10)') as geography,
+    STRUCT(1 AS a, 'abc' AS b) as s,
+    [1, 2, 3] as a,
+    [STRUCT(1 as a, ST_GEOGFROMTEXT('POINT (5 5)') as b), STRUCT(2, ST_GEOGFROMTEXT('POINT (10 10)')), STRUCT(3, ST_GEOGFROMTEXT('POINT (15 15)'))] as aos,
+    STRUCT([1, 2, 3] as a, ['a', 'b'] as b) as soa,
+    STRUCT([CAST ('Hi' as BYTES), CAST ('Bob' as BYTES)] as a, ['a', 'b'] as b) as bb,
+    STRUCT([ST_GEOGFROMTEXT('POINT (30 10)'), ST_GEOGFROMTEXT('POINT (15 15)')] as geo) as gg
+  FROM (SELECT DATETIME '2000-01-02 03:04:05.67' as datetime)"
+	tb <- bq_project_query(bq_test_project(), sql, quiet = TRUE)
+	df <- bqs_table_download(tb, bq_test_project(), as_tibble = TRUE, quiet = TRUE)
+	df2 <- bqs_table_download(tb, bq_test_project(), as_tibble = TRUE, quiet = TRUE, selected_fields = c("bb.a", "uniCODE", "aos"))
+  expect_equal(attr(df$datetime,"tzone"), "UTC")
+  expect_true(inherits(df$bytes, "blob"))
+  expect_true(inherits(df$geography, "wk_wkt"))
+  expect_true(inherits(df$aos[[1]]$b, "wk_wkt"))
+  expect_true(inherits(df$bb$a[[1]], "blob"))
+  expect_true(inherits(df$gg$geo[[1]], "wk_wkt"))
+  expect_equal(length(df2), 3)
 })