+++ NEWS.md
@@ -1,5 +1,8 @@
 # bigrquerystorage 1.1.0.9000
 
+* With `use_tibble = TRUE`
+  * Convert BigQuery SQL types BYTES, GEOGRAPHY to blob, wk_wkt
+  * Set timezone of BigQuery SQL type DATETIME to UTC
 * Fix nested list parse post processing.
 * Fix returning more rows than actual rows in source table when n_max > nrows.